- Introduce app/(medical)/ sub-navigator (Stack) with two screens:     MedicalRegisterList — shows all saved profiles with edit/delete       buttons and an \"Add\" button (disabled at MAX_REGISTERED_PERSONS)     MedicalRegister     — full form screen, now supports edit mode via       index route param; navigates back on save/cancel - app/MedicalRegister.tsx becomes a thin redirect to the list screen - All UI strings (confirm-delete, success/error alerts, btn labels)   added to lib/strings.ts; no hardcoded copy in screens"
